Rename celery factory function for clarity
